BBF-323 Improve direction facing for player
Also Add attribute to LFEffects to make them not play when already playing

+++ Assets/RedRat Productions/Scripts/Runtime/Core/Extensions/FloatExtensions.cs
@@ -11,7 +11,7 @@ namespace RedRats.Core
         /// <summary>
         /// Works like Mathf.Sign but with a 0.
         /// </summary>
-        public static float Sign0(this float number)
+        public static int Sign0(this float number)
         {
             return number == 0 ? 0 : number > 0 ? 1 : -1;
         }
